Add the 'no_root_squash' nfs option
This is required to perform the setfacl operations on an Ubuntu host.
On Arch this was not required (same NFS versions). It remains a mystery
why we need this on Ubuntu.

Usage of the 'no_root_squash' option is discouraged but in this isolated
dev environment it should not pose to much risk.